refactor(foundation): move ai_contract_diff types to foundation (A5)
Phase 6 / Sub-phase A5: Extract the 456-line ai_contract_diff/types.rs to
foundation. Zero crate deps (only chrono + serde + std). Used by:
- mockforge-recorder/sync_drift.rs (now migrated)
- mockforge-core/contract_drift (core shim remains)
- Every deprecated module that references ai_contract_diff types

Moved to mockforge-foundation::contract_diff_types:
- ContractDiffResult, Mismatch, MismatchType, MismatchSeverity
- Recommendation, CorrectionProposal, PatchOperation
- DiffMetadata, ContractDiffConfig, CapturedRequest, ConfidenceLevel

Migrated mockforge-recorder/src/sync_drift.rs to import Mismatch/MismatchSeverity/
MismatchType from foundation instead of core.

Foundation tests: 27 → 36 passing. Workspace builds clean.
